Add updated Fortran version of the program
arc_new.f is the non graphic version of the Fortran program with two
fixes/changes applied.

The bounds checking in the convection routine (lines 550-553) was fixed
so that out of bounds movement results in no change in temperatures.

The convection direction (lines 519-548) was changed to follow the
following direction matrix:

1  2  3
4  5  6
7  8  9

Also, the precision of the temperatures in the surfer file was changed.
The precision of the output temperatures can be changed on label 21
(line 618) where the format is:

FW.D
W = minimum width of the output
D = number of digits to the right of the decimal point