Added the initial C++ version of the finite difference heat flow simulation
arc.cpp is the initial C++ version of the finite difference heat flow
simulation with all fixes and changes described in the arc_new.f commit.

This version of the program should produce the same results as the
arc_new.f program with a small amount of deviation most likely caused by
rounding or the handling of REALs in Fortran.

It is formatted to run in the same manner as the non graphic Fortran
version of the program with similar output and prompts.