Updated the model to allow conduction to occur on the top (row 0 or depth 0), added additional save options, and updated function.
Updated the C++ and Fortran programs to allow conduction to occur at the
top of the model (row 0 or depth 0).

Added option to allow the user to save the final result without having
to save the models state at each screen update.

Removed unnecessary passing of global variable to perform_conv_2D
function.