Larger variable arrays, fixed title space problem, added input file conversion program, and added additional input files.
Extended the character width of indexes for convection and conduction
codes to two characters which increases the maximum number of possible
values from 9 to 99.

The hard coded constant arrays of size 8 are now variable in size with
their length defined in the input file.

Fixed problem where title's with multiple words would causes the
load_file function to crash due to misalignment in the input file.

Added program to convert old input files to the new format.

Converted the existing new input files to the new format and added
additional files.