Updated user input error checking, moving source, distance calculations. Added tolerance stopping condition. Fixed 3D input files.
The user input code was changed to handle incorrect input, such as a
character instead of an integer, and prevent infinite loops due to
incorrect input.

The moving source code was updated to allow for one or more rectangular
moving sources to be created where the user specifies the starting
location, size, velocity vector, acceleration, and temp.

Updated the distance calculations in the perform_convection function to
use the distance arrays used by the moving source.

Add the ability to have the program stop when a certain tolerance has
been met. The tolerance defines the minimum temperature change between
time steps for the program to continue running.

The number of slices in scale_3d.txt and uniform_4.txt was changed to 4.