Updated conduction codes, max temp for display purposes, moving source prompts, utility programs, and input files.
The conduction codes now include a 0 or 1 at the end of the code that
indicates if conduction can occur for the specified cell. The new input
files were updated to include this change.

All C++ files were changed from float to double precision.
The maximum temp used for display purposes is not updated each time the
display is updated.

The moving source prompts now let the user know the min and max values.

The utility programs were updated to prompt the user for input values
instead of command line arguments and to handle the new input file
format.

The floating point values used for display purposes are now typecasted
to GLfloat to remove warnings.